requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.36
pydantic==2.9.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3